floats this time

+++ payCalculator.py
@@ -4,8 +4,8 @@ def calculatePay():
     # This first line is provided for you
     hrs = input("Enter Hours:")
     rte = input("Enter rate of pay:")
-    pay = (int(hrs) * int(rte))
-    print(pay)
+    pay = (float(hrs) * float(rte))
+    print(float(pay))
 
     # end assignment
